added OpticalConf back into enums
