+++ python/lsst/ts/tunablelaser/csc.py
@@ -26,39 +26,12 @@ __all__ = ["run_tunablelaser", "LaserCSC"]
 
 import asyncio
 
-# TODO: (DM-46168) Revert workaround for TunableLaser XML changes
-import enum
-import warnings
-
 from lsst.ts import salobj, utils
 from lsst.ts.xml.enums.TunableLaser import LaserDetailedState, LaserErrorCode
 
-try:
-    from lsst.ts.xml.enums.TunableLaser import OpticalConfiguration
-except ImportError:
-    warnings.warn(
-        "OpticalConfiguration enumeration not availble in ts-xml. Using local version."
-    )
-
-    class OpticalConfiguration(enum.StrEnum):
-        """Configuration of the optical output"""
-
-        SCU = "SCU"
-        """Pass the beam straight-through the SCU."""
-        F1_SCU = "F1 SCU"
-        """Direct the beam through the F1 after passing through the SCU."""
-        F2_SCU = "F2 SCU"
-        """Direct the beam through the F2 after passing through the SCU."""
-        NO_SCU = "No SCU"
-        """Pass the beam straight-through."""
-        F1_NO_SCU = "F1 No SCU"
-        """Pass the beam to F1 output."""
-        F2_NO_SCU = "F2 No SCU"
-        """Pass the beam to F2 output."""
-
-
 from . import __version__, component, mock_server
 from .config_schema import CONFIG_SCHEMA
+from .enums import OpticalConfiguration
 
 
 def run_tunablelaser():
